코드 잔 수정 및 유효성 검사 추가

+++ server/controllers/webtoonDeleteController.js
@@ -5,8 +5,12 @@ const WebtoonDeleteController = {
     async deleteWebtoon(req, res) {
         try {
             const { EnName } = req.body;
-            await WebtoonDeleteService.deleteWebtoon(EnName);
-            res.send("삭제 성공");
+            if (!EnName) {
+                throw new Error('내용을 입력하세요');
+            }else{
+                const resultMessage = await WebtoonDeleteService.deleteWebtoon(EnName);
+                res.send(resultMessage);
+            }
         } catch (error) {
             console.error(error);
             res.status(500).json({ message: '서버 오류' });
@@ -17,8 +21,12 @@ const WebtoonDeleteController = {
     async deleteEpisode(req, res) {
         try {
             const { EnName, ep } = req.body;
-            await WebtoonDeleteService.deleteEpisode( EnName, ep);
-            res.send(ep+"화 삭제 성공");
+            if (!EnName || !ep) {
+                throw new Error('내용을 입력하세요');
+            }else{
+                const resultMessage = await WebtoonDeleteService.deleteEpisode( EnName, ep);
+                res.send(resultMessage);
+            }
         } catch (error) {
             console.error(error);
             res.status(500).json({ message: '서버 오류' });
